refactor(Rust): Rename `graph`and create `graph-triples` crates
