+++ node/src/graphs.rs
@@ -4,6 +4,6 @@ use stencila::graphs;
 
 /// Get the module's schemas
 pub fn schemas(cx: FunctionContext) -> JsResult<JsString> {
-    let schemas = graphs::schemas();
+    let schemas = graph::schemas();
     to_json_or_throw(cx, schemas)
 }
