+++ node/src/lib.rs
@@ -49,7 +49,7 @@ fn main(mut cx: ModuleContext) -> NeonResult<()> {
     cx.export_function("formatsSchemas", formats::schemas)?;
     cx.export_function("formatsFormats", formats::formats)?;
 
-    cx.export_function("graphsSchemas", graphs::schemas)?;
+    cx.export_function("graphsSchemas", graph::schemas)?;
 
     cx.export_function("sourcesSchemas", sources::schemas)?;
