+++ rust/graph-triples/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "graph-triples"
+version = "0.0.0"
+edition = "2021"
+
+[dependencies]
+derivative = "2.2.0"
+eyre = "0.6.5"
+path-slash = "0.1.4"
+petgraph = { version = "0.6.0", features = ["serde-1"] }
+schemars = { version = "0.8.6", features = ["preserve_order"] }
+serde = "1.0.130"
+serde_json = { version = "1.0.68", features = ["preserve_order"] }
+serde_with = "1.11.0"
+strum = { version = "0.22.0", features = ["derive"] }
+utils = { path = "../utils", version = "0.0.0" }
