+++ rust/graph-triples/src/directions.rs
@@ -1,0 +1,22 @@
+use crate::Relation;
+
+/// The direction to represent the flow of information from subject to object
+pub enum Direction {
+    From,
+    To,
+}
+
+/// Get the the `Direction` for a `Relation`
+pub fn direction(relation: &Relation) -> Direction {
+    match relation {
+        Relation::Assign(..) => Direction::To,
+        Relation::Convert(..) => Direction::To,
+        Relation::Embed => Direction::From,
+        Relation::Import(..) => Direction::To,
+        Relation::Include => Direction::From,
+        Relation::Link => Direction::To,
+        Relation::Read(..) => Direction::From,
+        Relation::Use(..) => Direction::From,
+        Relation::Write(..) => Direction::To,
+    }
+}
