+++ rust/graph/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
-name = "graphs"
+name = "graph"
 version = "0.0.0"
 edition = "2021"
 
 [dependencies]
 derivative = "2.2.0"
 eyre = "0.6.5"
+graph-triples = { path = "../graph-triples", version = "0.0.0" }
 path-slash = "0.1.4"
 petgraph = { version = "0.6.0", features = ["serde-1"] }
 schemars = { version = "0.8.6", features = ["preserve_order"] }
