+++ rust/graph/src/lib.rs
@@ -1,5 +1,5 @@
-use derivative::Derivative;
 use eyre::Result;
+use graph_triples::{direction, relations, Direction, Relation, Resource, Triple};
 use path_slash::PathExt;
 use petgraph::{
     graph::NodeIndex,
@@ -20,284 +20,6 @@ use std::{
 };
 use strum::Display;
 use utils::some_string;
-
-/// A resource in a dependency graph (the nodes of the graph)
-#[derive(Debug, Clone, PartialEq, Eq, Hash, JsonSchema, Serialize)]
-#[serde(tag = "type")]
-pub enum Resource {
-    /// A symbol within code, within a project file
-    Symbol(resources::Symbol),
-
-    /// A node within a project file
-    Node(resources::Node),
-
-    /// A file within the project
-    File(resources::File),
-
-    /// A declared project `Source`
-    Source(resources::Source),
-
-    /// A programming language module, usually part of an external package
-    Module(resources::Module),
-
-    /// A URL to a remote resource
-    Url(resources::Url),
-}
-
-pub mod resources {
-    use super::*;
-    use std::path::{Path, PathBuf};
-
-    #[derive(Debug, Clone, Derivative, JsonSchema, Serialize)]
-    #[derivative(PartialEq, Eq, Hash)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Symbol {
-        /// The path of the file that the symbol is defined in
-        #[serde(serialize_with = "serialize_path")]
-        pub path: PathBuf,
-
-        /// The name/identifier of the symbol
-        pub name: String,
-
-        /// The type of the object that the symbol refers to (e.g `Number`, `Function`)
-        ///
-        /// Should be used as a hint only, and as such is excluded from
-        /// equality and hash functions.
-        #[derivative(PartialEq = "ignore")]
-        #[derivative(Hash = "ignore")]
-        pub kind: String,
-    }
-
-    /// Create a new `Symbol` resource
-    pub fn symbol(path: &Path, name: &str, kind: &str) -> Resource {
-        Resource::Symbol(Symbol {
-            path: path.to_path_buf(),
-            name: name.into(),
-            kind: kind.into(),
-        })
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, Hash, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Node {
-        /// The path of the file that the node is defined in
-        #[serde(serialize_with = "serialize_path")]
-        pub path: PathBuf,
-
-        /// The id of the node with the document
-        pub id: String,
-
-        /// The type of node e.g. `Parameter`, `CodeChunk`
-        pub kind: String,
-    }
-
-    /// Create a new `Node` resource
-    pub fn node(path: &Path, id: &str, kind: &str) -> Resource {
-        Resource::Node(Node {
-            path: path.to_path_buf(),
-            id: id.into(),
-            kind: kind.into(),
-        })
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, Hash, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct File {
-        /// The path of the file
-        #[serde(serialize_with = "serialize_path")]
-        pub path: PathBuf,
-    }
-
-    /// Create a new `File` resource
-    pub fn file(path: &Path) -> Resource {
-        Resource::File(File {
-            path: path.to_path_buf(),
-        })
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, Hash, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Source {
-        /// The name of the project source
-        pub name: String,
-    }
-
-    /// Create a new `Source` resource
-    pub fn source(name: &str) -> Resource {
-        Resource::Source(Source { name: name.into() })
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, Hash, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Module {
-        /// The programming language of the module
-        pub language: String,
-
-        /// The name of the module
-        pub name: String,
-    }
-
-    /// Create a new `Module` resource
-    pub fn module(language: &str, name: &str) -> Resource {
-        Resource::Module(Module {
-            language: language.into(),
-            name: name.into(),
-        })
-    }
-
-    #[derive(Debug, Clone, Default, PartialEq, Eq, Hash, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Url {
-        /// The URL of the external resource
-        pub url: String,
-    }
-
-    /// Create a new `Url` resource
-    pub fn url(url: &str) -> Resource {
-        Resource::Url(Url { url: url.into() })
-    }
-}
-
-/// Serialize the `path` fields of resources so that they use Unix forward slash
-/// separators on all platforms.
-fn serialize_path<S>(path: &Path, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    path.to_slash_lossy().serialize(serializer)
-}
-
-/// The relation between two resources in a dependency graph (the edges of the graph)
-///
-/// Some relations carry additional information such whether the relation is active
-/// (`Import` and `Convert`) or the range that they occur in code (`Assign`, `Use`, `Read`) etc
-#[derive(Debug, Display, Clone, JsonSchema, Serialize)]
-#[serde(tag = "type")]
-pub enum Relation {
-    Assign(relations::Assign),
-    Convert(relations::Convert),
-    Embed,
-    Import(relations::Import),
-    Include,
-    Link,
-    Read(relations::Read),
-    Use(relations::Use),
-    Write(relations::Write),
-}
-
-/// The two dimensional range that a relation is defined within some
-/// code (line start, column start, line end, column end).
-pub type Range = (usize, usize, usize, usize);
-
-/// A null range which can be used in places where we do not know where
-/// in the `subject` the relation is defined.
-pub const NULL_RANGE: Range = (0, 0, 0, 0);
-
-pub mod relations {
-    use super::*;
-
-    /// Assigns a symbol
-    #[derive(Debug, Clone, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Assign {
-        /// The range within code that the assignment is done
-        pub range: Range,
-    }
-
-    /// Create a new `Assign` relation
-    pub fn assigns(range: Range) -> Relation {
-        Relation::Assign(Assign { range })
-    }
-
-    /// Imports a file from a `Source`
-    #[derive(Debug, Clone, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Import {
-        /// Whether or not the import is automatically updated
-        pub auto: bool,
-    }
-
-    /// Create a new `Import` relation
-    pub fn imports(auto: bool) -> Relation {
-        Relation::Import(Import { auto })
-    }
-
-    /// Converts a file into another
-    #[derive(Debug, Clone, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Convert {
-        /// Whether or not the conversion is automatically updated
-        pub auto: bool,
-    }
-
-    /// Create a new `Convert` relation
-    pub fn converts(auto: bool) -> Relation {
-        Relation::Convert(Convert { auto })
-    }
-
-    /// Reads from a file
-    #[derive(Debug, Clone, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Read {
-        /// The range within code that the read is declared
-        pub range: Range,
-    }
-
-    /// Create a new `Read` relation
-    pub fn reads(range: Range) -> Relation {
-        Relation::Read(Read { range })
-    }
-
-    /// Uses a symbol or module
-    #[derive(Debug, Clone, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Use {
-        /// The range within code that the use is declared
-        pub range: Range,
-    }
-
-    /// Create a new `Use` relation
-    pub fn uses(range: Range) -> Relation {
-        Relation::Use(Use { range })
-    }
-
-    /// Writes to a file
-    #[derive(Debug, Clone, JsonSchema, Serialize)]
-    #[schemars(deny_unknown_fields)]
-    pub struct Write {
-        /// The range within code that the write is declared
-        pub range: Range,
-    }
-
-    /// Create a new `Write` relation
-    pub fn writes(range: Range) -> Relation {
-        Relation::Write(Write { range })
-    }
-}
-
-/// The direction to represent the flow of information from subject to object
-pub enum Direction {
-    From,
-    To,
-}
-
-/// Get the the `Direction` for a `Relation`
-pub fn direction(relation: &Relation) -> Direction {
-    match relation {
-        Relation::Assign(..) => Direction::To,
-        Relation::Convert(..) => Direction::To,
-        Relation::Embed => Direction::From,
-        Relation::Import(..) => Direction::To,
-        Relation::Include => Direction::From,
-        Relation::Link => Direction::To,
-        Relation::Read(..) => Direction::From,
-        Relation::Use(..) => Direction::From,
-        Relation::Write(..) => Direction::To,
-    }
-}
-
-/// A subject-relation-object triple
-pub type Triple = (Resource, Relation, Resource);
 
 /// A project dependency graph
 #[derive(Debug, Default, Clone)]
@@ -547,7 +269,7 @@ impl Graph {
                         | Relation::Use(relations::Use { range })
                         | Relation::Read(relations::Read { range })
                         | Relation::Write(relations::Write { range }) => {
-                            let label = if *range == NULL_RANGE {
+                            let label = if *range == relations::NULL_RANGE {
                                 relation.to_string()
                             } else {
                                 format!("{} L{}", relation, range.0 + 1)
