+++ rust/stencila/Cargo.toml
@@ -201,7 +201,7 @@ codec-ipynb = { path = "../codec-ipynb", version = "0.0.0" }
 codec-txt = { path = "../codec-txt", version = "0.0.0" }
 codecs = { path = "../codecs", version = "0.0.0" }
 formats = { path = "../formats", version = "0.0.0" }
-graphs = { path = "../graphs", version = "0.0.0" }
+graph = { path = "../graph", version = "0.0.0" }
 
 [dev-dependencies]
 criterion = "0.3.5"
