+++ rust/stencila/src/documents.rs
@@ -13,7 +13,7 @@ use crate::{
 use defaults::Defaults;
 use eyre::{bail, Result};
 use formats::{Format, FORMATS};
-use graphs::{Relation, Resource};
+use graph::{Relation, Resource};
 use itertools::Itertools;
 use maplit::hashset;
 use notify::DebouncedEvent;
