+++ rust/stencila/src/kernels/mod.rs
@@ -4,7 +4,7 @@ use chrono::{DateTime, Utc};
 use derive_more::{Deref, DerefMut};
 use enum_dispatch::enum_dispatch;
 use eyre::{eyre, Result};
-use graphs::{Relation, Resource};
+use graph::{Relation, Resource};
 use schemars::JsonSchema;
 use serde::Serialize;
 use std::collections::{hash_map::Entry, BTreeMap, HashMap};
