+++ rust/stencila/src/methods/compile/code/ts.rs
@@ -2,7 +2,7 @@ use super::{
     helpers::{apply_tags, Compiler},
     js::{self, handle_patterns},
 };
-use graphs::{relations, resources, Relation, Resource};
+use graph::{relations, resources, Relation, Resource};
 use once_cell::sync::Lazy;
 use std::path::Path;
