+++ rust/stencila/src/projects.rs
@@ -8,7 +8,7 @@ use crate::pubsub::publish;
 use crate::sources::{self, Source, SourceDestination, SourceTrait};
 use crate::utils::schemas;
 use eyre::{bail, Result};
-use graphs::{resources, Graph, GraphEvent, GraphEventType, Resource};
+use graph::{resources, Graph, GraphEvent, GraphEventType, Resource};
 use notify::DebouncedEvent;
 use once_cell::sync::Lazy;
 use regex::Regex;
